\name{pbListPosts}
\alias{pbListPosts}
\title{List Pastebin posts from a selected user}
\description{
  Lists the unique paste IDs listed on a user's pastebin account.
}
\usage{
  pbListPosts(user = "beepbeepbeep", api_results_limit = 1000, session_key)
}
\arguments{
  \item{user}{The user whose posts you'd like a list of. Troy Hill's username ("beepbeepbeep") is used by default}
  
  \item{api_results_limit}{Number of results to list (must be between zero and 1000)}

  \item{session_key}{Session key, generated by /code{getUserKey}}
}
}
\value{
  \item{paste IDs}{a vector containing the paste IDs (the unique portion of the web address, following 'pastebin.com/') of the user's public posts (or all posts, if the targeted account is your own)}
}
\examples{
# generate a session key first
# a <- pbListPosts(session_key = key) # lists all of beepbeepbeep's posts
}