\name{pbPost}
\alias{pbPost}
\title{Post text to Pastebin}
\description{
  Posts text to pastebin as a guest or as a user.
}
\usage{
  pbPost(session_key = "", api_dev_key = "", text = "hello world!", title = "Untitled", expire = "1M", privacy_code = "0")
}
\arguments{
  \item{session_key}{Session key, generated by /code{getUserKey}}
  
  \item{api_dev_key}{Development key for Pastebin API, from http://pastebin.com/api/}
  
  \item{text}{Text or vector of elements to be posted. For vectors with mutliple 
  elements, a paste will be created for each element.}
  
  \item{title}{Title of paste (untitled by default).}

  \item{expire}{Time until post expires. Options are ten minutes, one hour,
  one day, one week, 2 weeks, one month, or never (respectively: "N", 
  "10M", "1H", "1D", "1W", "2W", "1M")}
  
  \item{privacy_code}{Whether post is public (default), private, or unlisted ("0", "1", or "2", respectively)}
}
\value{
  \item{pasteID}{The web address of the generated paste(s).}
}
\examples{
# text_items <- c("hello world!", "hi there, world!")
# posts <- pbPost(text = text_items, session_key = "Your_session_key_goes_here", expire = "1M")
# posts
# 
# x = runif(150)*100
# pbPost(text = as.character(dput(x)), session_key = "Your_session_key_goes_here", expire = "1M")
}