\name{pbPost}
\alias{pbPost}
\title{Post text to Pastebin}
\description{
  Posts text to pastebin as a guest or as a user.
}
\usage{
  pbPost(text = "Hello world!", title = "Untitled", 
                   user_key = "", expire = "1W")
}
\arguments{
  \item{text}{Text or vector of elements to be posted. For vectors with mutliple 
  elements, a paste will be created for each element.}
  
  \item{title}{Title of paste (untitled by default).}

  \item{user_key}{Session key, generated by /code{getUserKey}}
  
  \item{expire}{Time until post expires. Options are ten minutes, one hour,
  one day, one week, 2 weeks, one month, or never (respectively: "N", 
  "10M", "1H", "1D", "1W", "2W", "1M")}
}
\value{
  \item{pasteID}{The web address of the generated paste(s).}
}
\examples{
# text_items <- c("hello world!", "hi there, world!")
# posts <- pbPost(text = text_items, user_key = api_user_key, expire = "1M")
# posts
# 
# x = runif(150)*100
# pbPost(text = as.character(dput(x)), user_key = api_user_key, expire = "1M")
}